Share types between builder and template
