+++ bin/elytre.js
@@ -1,6 +1,6 @@
 #!/usr/bin/env node
 
-const build = require('../builder/dist/index.js').default;
+const build = require('../builder/dist/builder/src/build').default;
 
 const [command] = process.argv.slice(2);
