+++ builder/src/__mocks__/build-catalog.ts
@@ -1,4 +1,4 @@
-import { Catalog } from '../types';
+import { Catalog } from '../../../shared/types';
 
 export default function buildCatalog(): Catalog {
   return {
