+++ builder/src/process-covers.ts
@@ -1,8 +1,8 @@
 import chalk from 'chalk';
 import { existsSync, readdirSync, copyFileSync } from 'fs-extra';
-import log from './log';
 
-import { Product } from './types';
+import log from './log';
+import { Product } from '../../shared/types';
 
 export default function processCovers(
   products: Product[],
