+++ template/src/components/ProductList.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 
-import { Product as ProductType } from '../lib/user-files';
+import { Product as ProductType } from '../../../shared/types';
 import Product from './Product';
 
 type ProductListProps = {
