--- template/src/shared/types.ts
@@ -1,1 +1,0 @@
-export type Locale = 'en' | 'fr';
